Add CreateSSOConnection
