+++ components/setup-link-instructions/auth0/4.mdx
@@ -1,5 +1,5 @@
 ## Step 4: Create SAML Connection
 
-<CreateConnection setupLinkToken={token} idpEntityID={idpEntityId} />
+<CreateSSOConnection setupLinkToken={setupLinkToken} idpType='saml' />
 
 <Footer hasNext={false} />
