+++ components/setup-link-instructions/okta/4.mdx
@@ -1,5 +1,5 @@
 ## Step 4: Create SAML Connection
 
-<CreateSAMLConnection setupLinkToken={token} classNames={{}} />
+<CreateSSOConnection setupLinkToken={setupLinkToken} idpType='saml' />
 
 <Footer hasNext={false} />
